pyproject.toml
[tool.poetry]
name = "recommender"
version = "0.1.0"
description = ""
authors = ["Mark"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohappyeyeballs = "2.4.0"
aiohttp = "3.10.5"
aiosignal = "1.3.1"
alembic = "1.13.2"
annotated-types = "0.7.0"
anyio = "4.6.0"
async-timeout = "4.0.3"
attrs = "24.2.0"
blinker = "1.8.2"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
distro = "1.9.0"
exceptiongroup = "1.2.2"
filelock = "3.16.1"
flask = "3.0.3"
flask-cors = "5.0.0"
flask-migrate = "4.0.7"
flask-sqlalchemy = "3.1.1"
frozenlist = "1.4.1"
fsspec = "2024.10.0"
greenlet = "3.1.0"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.2"
idna = "3.10"
instagrapi = "2.1.2"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
jiter = "0.5.0"
llvmlite = "0.43.0"
mako = "1.3.5"
markupsafe = "2.1.5"
more-itertools = "10.5.0"
mpmath = "1.3.0"
multidict = "6.1.0"
networkx = "3.4.2"
numba = "0.60.0"
numpy = "2.0.2"
openai = "0.28.0"
openai-whisper = "20240930"
outcome = "1.3.0.post0"
packaging = "24.1"
pillow = "10.4.0"
psycopg2-binary = "2.9.9"
pycparser = "2.22"
pycryptodomex = "3.20.0"
pydantic = "2.7.1"
pydantic-core = "2.18.2"
pydub = "0.25.1"
pyperclip = "1.9.0"
pysocks = "1.7.1"
python-dotenv = "1.0.1"
pytz = "2024.2"
rapidfuzz = "3.10.0"
regex = "2024.9.11"
requests = "2.32.3"
selenium = "4.25.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
speechrecognition = "3.11.0"
sqlalchemy = "2.0.35"
sympy = "1.13.1"
thefuzz = "0.22.1"
tiktoken = "0.7.0"
torch = "2.5.0"
tqdm = "4.66.5"
trio = "0.27.0"
trio-websocket = "0.11.1"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
webdriver-manager = "4.0.2"
websocket-client = "1.8.0"
werkzeug = "3.0.4"
wsproto = "1.2.0"
yarl = "1.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
